+++ public/utils/check-plugin-version.ts
@@ -9,61 +9,64 @@
  *
  * Find more information about this on the LICENSE file.
  */
-import { GenericRequest } from "../react-services/generic-request";
+import { GenericRequest } from '../react-services/generic-request';
 import { AxiosResponse } from 'axios';
-import _ from "lodash";
+import _ from 'lodash';
+import meta from '../assets/meta.json';
+import { getCookies } from '../kibana-services';
 
 type TAppInfo = {
-  revision: string
-  'app-version': string
-}
+  revision: string;
+  'app-version': string;
+};
 
 type TAppInfoResponse = {
-  statusCode: number
-  data: TAppInfo
-}
-
-type cookie = [string, string];
-
-const wazuhCookies: cookie[] = [
-  ["currentApi", "/app/wazuh"],
-  ["APISelector", "/app/wazuh"],
-  ["clusterInfo", "/app/wazuh"],
-  ["currentPattern", "/app/wazuh"],
-  ["patternSelector", "/app/wazuh"],
-]
+  statusCode: number;
+  data: TAppInfo;
+};
 
 export const checkPluginVersion = async () => {
   try {
-    const response: AxiosResponse<TAppInfoResponse> = await GenericRequest.request('GET', '/api/setup');
-    const { revision, "app-version": appRevision } = response.data.data;
-    return checkLocalstorageVersion({ revision, "app-version": appRevision });
+    const response: AxiosResponse<TAppInfoResponse> = await GenericRequest.request(
+      'GET',
+      '/api/setup'
+    );
+    const { revision, 'app-version': appRevision } = response.data.data;
+    return checkClientAppVersion({ revision, 'app-version': appRevision });
   } catch (error) {
-    console.error(`Error when getting the plugin version: ${error}`)
+    console.error(`Error when getting the plugin version: ${error}`);
   }
-}
+};
 
-const checkLocalstorageVersion = (appInfo: TAppInfo) => {
-  const storeAppInfo = localStorage.getItem('appInfo');
-
-  if (!storeAppInfo) {
-    updateAppInfo(appInfo);
-    return;
+const checkClientAppVersion = (appInfo: TAppInfo) => {
+  if (appInfo['app-version'] !== meta.version || appInfo.revision !== meta.revision) {
+    clearBrowserInfo(appInfo);
+  } else {
+    const storeAppInfo = localStorage.getItem('appInfo');
+    !storeAppInfo && updateAppInfo(appInfo);
   }
-
-  !_.isEqual(appInfo, JSON.parse(storeAppInfo)) && clearBrowserInfo(appInfo);
-}
-
-const deleteWazuhCookies = ([name, path]: cookie) => {
-  document.cookie = `${name}=; expires=Thu, 01 Jan 1970 00:00:00 UTC; path=${path};`;
-}
+};
 
 function clearBrowserInfo(appInfo: TAppInfo) {
-  wazuhCookies.forEach(deleteWazuhCookies);
+  console.warn('Clearing browser cache');
+  //remove cookies
+  const cookies = getCookies().getAll();
+  Object.keys(cookies).forEach((cookie) => getCookies().remove(cookie));
+
+  //remove cache
+  if (caches) {    
+    caches.keys().then(function (names) {
+      for (let name of names) caches.delete(name);
+    });
+  }
+
+  //update localStorage
   updateAppInfo(appInfo);
+
+  // delete browser cache and hard reload
+  window.location.reload(true);
 }
 
 function updateAppInfo(appInfo: TAppInfo) {
-  localStorage.setItem("appInfo", JSON.stringify(appInfo));
+  localStorage.setItem('appInfo', JSON.stringify(appInfo));
 }
-