+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 
 All notable changes to the Wazuh app project will be documented in this file.
+
+## Wazuh v4.1.2 - Kibana 7.10.0 , 7.10.2 - Revision 4103
+
+### Fixed
+
+- Improved validation and prevention for caching bundles in client side []()
+
 ## Wazuh v4.1.2 - Kibana 7.10.0 , 7.10.2 - Revision 4103
 
 ### Added